moved sample data to separate module so it can be used in both rest and web module
