+++ pkmnleague-data/pom.xml
@@ -8,10 +8,9 @@
         <version>1.0-SNAPSHOT</version>
     </parent>
 
-    <artifactId>pkmnleague-web</artifactId>
-    <packaging>war</packaging>
+    <artifactId>pkmnleague-data</artifactId>
 
-    <name>Web Module: Frontend</name>
+    <name>Sample data</name>
 
     <dependencies>
         <dependency>
@@ -89,11 +88,6 @@
             <groupId>org.mockito</groupId>
             <artifactId>mockito-all</artifactId>
             <scope>test</scope>
-        </dependency>
-        <dependency>
-            <groupId>cz.fi.muni.pa165.seminar.pkmnleague</groupId>
-            <artifactId>pkmnleague-data</artifactId>
-            <version>1.0-SNAPSHOT</version>
         </dependency>
     </dependencies>
