+++ pom.xml
@@ -45,6 +45,7 @@
         <module>pkmnleague-service</module>
         <module>pkmnleague-api</module>
         <module>pkmnleague-web</module>
+        <module>pkmnleague-data</module>
         <module>pkmnleague-rest</module>
     </modules>
 </project>
